third_party/blink/web_tests/external/wpt/html/infrastructure/urls/resolving-urls/query-encoding/windows-1252_include=loading-expected.txt

This is a testharness.js-based test.
[FAIL] loading image <embed src>
  assert_equals: expected substring %E5 got default intrinsic width expected 1 but got 300
[FAIL] loading image <object data>
  assert_equals: expected substring %E5 got default intrinsic width expected 1 but got 300
Harness: the test ran to completion.
